fix: norminette + extra files
